acp: rename recent-sessions button to ACPSessionsButton
Icon-only button now absorbs the session title, rename, and
disconnected/reconnect surfaces that previously lived inline on
ACPToolbar. Red dot on the icon signals disconnected at a glance.
Build is broken until the toolbar call site updates in the next
commit.
